Add files via upload
This is the ER diagram of the University displayed as an example to display various entities along with different types of attributes. The Diagram contains "5" main entities which are as follows - STUDENT, DEPARTMENT, FACULTY, PROJECT, and SUBJECT. each entity has its own attributes and role of each special attribute is denoted with the appropriate diagrammatic notation.